tests: fix prctl marker
On some platforms, unsigned int arguments passed to syscall() are sign-extended.
Workaround this by casting prctl marker arguments to unsigned long.

* tests/prctl-arg2-intptr.c (main): Cast last four prctl marker
arguments to unsigned long.
* tests/prctl-cap-ambient.c (main): Likewise.
* tests/prctl-capbset.c (main): Likewise.
* tests/prctl-dumpable.c (main): Likewise.
* tests/prctl-fp-mode.c (main): Likewise.
* tests/prctl-mce-kill.c (main): Likewise.
* tests/prctl-name.c (main): Likewise.
* tests/prctl-no-args.c (main): Likewise.
* tests/prctl-pdeathsig.c (main): Likewise.
* tests/prctl-seccomp-filter-v.c (main): Likewise.
* tests/prctl-seccomp-strict.c (main): Likewise.
* tests/prctl-securebits.c (prctl): Likewise.
* tests/prctl-spec-inject.c (main): Likewise.
* tests/prctl-sve.c (main): Likewise.
* tests/prctl-tid_address.c (main): Likewise.
* tests/prctl-tsc.c (main): Likewise.
* tests/prctl-unalign.c (main): Likewise.

Fixes: v5.12-310-gb9f9d5b25 "tests: unify common code in prctl decoder tests"
Fixes: v5.12-311-gcfadf5f5a "tests: improve test coverage of prctl syscall decoder"

+++ tests/prctl-arg2-intptr.c
@@ -53,7 +53,8 @@ main(void)
 	long rc;
 	unsigned int i;
 
-	syscall(__NR_prctl, -1U, -2U, -3U, -4U, -5U);
+	syscall(__NR_prctl, -1U, (unsigned long) -2U, (unsigned long) -3U,
+				 (unsigned long) -4U, (unsigned long) -5U);
 
 	for (i = 0; i < ARRAY_SIZE(options); ++i) {
 		prctl(options[i].val | bogus_op_bits, 0);
