+++ tests/prctl-name.c
@@ -19,7 +19,8 @@
 int
 main(void)
 {
-	syscall(__NR_prctl, -1U, -2U, -3U, -4U, -5U);
+	syscall(__NR_prctl, -1U, (unsigned long) -2U, (unsigned long) -3U,
+				 (unsigned long) -4U, (unsigned long) -5U);
 
 	static const char str[] = "0123456789abcdef";
 	static const int len = sizeof(str) - 1;
